Extract Aspire.Hosting.Azure.Tests
Contributes to #4294

This also allows us to remove the Aspire.Hosting.Tests.SharedShim, now that VolumeNameGenerator is only found in Aspire.Hosting.
