+++ src/Aspire.Hosting.Azure.Search/Aspire.Hosting.Azure.Search.csproj
@@ -18,8 +18,4 @@
     <PackageReference Include="Azure.Provisioning.Search" />
   </ItemGroup>
 
-  <ItemGroup>
-    <InternalsVisibleTo Include="Aspire.Hosting.Tests" />
-  </ItemGroup>
-
 </Project>
