+++ src/Aspire.Hosting.Azure.SignalR/Aspire.Hosting.Azure.SignalR.csproj
@@ -18,8 +18,4 @@
     <PackageReference Include="Azure.Provisioning.SignalR" />
   </ItemGroup>
 
-  <ItemGroup>
-    <InternalsVisibleTo Include="Aspire.Hosting.Tests" />
-  </ItemGroup>
-
 </Project>
