+++ tests/Aspire.Hosting.Tests/Utils/VolumeNameGeneratorTests.cs
@@ -1,9 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-extern alias AspireHostingShared;
-
-using static AspireHostingShared::Aspire.Hosting.Utils.VolumeNameGenerator;
+using static Aspire.Hosting.Utils.VolumeNameGenerator;
 using Xunit;
 
 namespace Aspire.Hosting.Tests.Utils;
